Jormungandr: change parser of from_datetime of Schedules
